Smoke: move nydusd tool code into tests
