+++ tests/nydusd.go
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package tool
+package tests
 
 import (
 	"bytes"
